requirements.txt
Django==3.2.10
mysqlclient==2.1.0
djangorestframework==3.12.2
django-cors-headers==3.5.0
django-mysql==3.9